refactor: move the types to separate files
