+++ src/components/queue-page/types.ts
@@ -1,0 +1,5 @@
+export interface IQueue<T> {
+  enqueue: (item: T) => void;
+  dequeue: () => void;
+  clear: () => void;
+}
