+++ src/components/stack-page/types.ts
@@ -1,0 +1,5 @@
+export interface IStack<T> {
+  push: (item: T) => void;
+  pop: () => void;
+  clear: () => void;
+}
